\name{play}
\alias{play}
\title{Creates a spinning cube!}
\description{
  Initiates the rotation of the cube. A new window opens with a cube rotating on your screen!
}
\usage{
  play()
}
\value{
  A spinning cube.
}
\examples{
  # Rotate the cube with the following command:

  play()
  
  # A window with a new R Graphics Device appears with the rotating cube.
  
  # To interrupt the rotation:
  #   • On Windows: just press the red "STOP" button of your R GUI or RStudio and close the window.
  #   • On Linux: If you are using the terminal, just press Ctrl + C.
}  
